xivgear: handle relic stats correctly

+++ src/model/materia.rs
@@ -3,6 +3,6 @@ pub struct Materia {
     pub id: u32,
 
     pub item_id: Vec<u32>,
-    pub base_param_id: u32,
-    pub base_param_value: Vec<u32>,
+    pub base_param_id: i32,
+    pub base_param_value: Vec<i16>,
 }
